refactoring code for readability and adding a script for the database creation
