+++ readme.md
@@ -1,3 +1,11 @@
 This application is designed to pull in Pokemon stats from an API
 
 The purpose of this project is to utilize AWS services while incorporating devops practices.
+
+
+To utilize this project you need the following:
+
+- MYSQL
+- Install these python packages: mysql-connector, flask, pandas, requests, pip install sqlalchemy
+
+Var.py file. In this file create the variables db_user, db_password, and host. For this project host is localhost. Your db credentials will be defined when you set up your mysql db
