+++ scripts/insert_poke_sprites.py
@@ -6,6 +6,8 @@ import sqlite3
 from sqlalchemy import create_engine
 from datetime import datetime
 import json
+from var import *
+
 
 def get_sprite(x):
         while True:
